Filter Covariates by Removing Linearly Dependent Principal Components

This function filters covariates by removing linearly dependent principal components (PCs) from a genotype matrix (X) and a covariate matrix (C). The number of PCs to consider can either be specified directly or determined using the `varPCs` function. It performs Principal Component Analysis (PCA) on the genotype data and checks for linear dependence in the combined matrix of covariates and PCs.

coVar_Filter(X, C, num_PCs = 10, variance_threshold = 0.80, remove_NAs = FALSE)

Arguments

X

A numeric matrix or data frame representing the genotype data (SNPs). The rows represent individuals, and columns represent SNPs. The first column, if non-numeric, will be removed assuming it's an ID column.

C

A numeric matrix or data frame representing covariates. The first column, if non-numeric, will be removed assuming it's an ID column.

num_PCs

Either a numeric value indicating the number of principal components to retain or a character string `"varPCs"`. If `"varPCs"`, the function uses the `varPCs` function to determine the number of PCs that explain the specified variance threshold. Default is 10.

variance_threshold

A numeric value between 0 and 1 indicating the proportion of the variance to explain using the principal components. Only used if `num_PCs = "varPCs"`. Default is 0.80 (80

remove_NAs

A logical value. If `TRUE`, rows containing missing values (NA) in either `X` or `C` will be removed. If `FALSE`, the function will stop if missing values are detected. Default is `FALSE`.

Details

This function combines the PCs obtained from the genotype data (`X`) with the covariate matrix (`C`), then removes linearly dependent columns using linear regression to detect dependence. The returned matrix contains only the covariates and PCs that are not linearly dependent on each other.

Value

A numeric matrix or data frame of covariates (C) with linearly dependent PCs removed. The returned matrix includes both the original covariates and the retained PCs.

References

See the `varPCs` function for more details on determining the number of principal components to retain.

Author

Michelle Marcano-Delgado

Note

This function is useful for filtering out PCs that may not provide independent variance, ensuring that the remaining PCs are orthogonal to each other.

See also

Examples

## Example 1: Filter covariates by removing linearly dependent PCs (keeping top 10 PCs)
 set.seed(123)
genotype_matrix <- matrix(sample(c(0, 1, 2), 100 * 500, replace = TRUE), ncol = 500)  # Genotype matrix
covariate_matrix <- matrix(rnorm(500 * 5), ncol = 5)  # Covariates

filtered_covariates <- coVar_Filter(genotype_matrix, covariate_matrix, num_PCs = 10)
head(filtered_covariates)
#>            1            2          3             4          5        PC1
#> 1 -0.3202214  0.005074796 -0.1551520 -0.2988862816 -0.5468950  4.7236925
#> 2  0.2274341  1.611618513 -0.5453526 -0.0006212909 -1.2365944  3.7276472
#> 3  0.1948882  0.509342670 -1.2792133  2.4435849328  0.2523382  1.7326849
#> 4  0.1698660  1.080393772  0.1379318  0.8393081130 -0.3356361  0.7628494
#> 5 -0.4889679  1.312839587  1.0734213  0.0640024830 -0.6295072 -2.3226955
#> 6 -1.2607975 -0.858659339 -0.5883604  0.0848489681  0.1339525  3.7654215
#>          PC2        PC3          PC4        PC5        PC6        PC7
#> 1  0.4098598 -0.2887493  2.293134606 -3.3197207  2.6069942 -1.4108729
#> 2 -0.4531534 -2.9462645  0.008275828  2.5126370 -1.1261136 -1.1726880
#> 3 -1.7341891 -1.9878724 -0.253216763 -3.8908908 -4.6750506  0.5539815
#> 4 -6.1098026  2.6841439  2.860310716  1.3495148  0.2447669  1.2634274
#> 5  0.8032023 -5.1038478  0.978313352  0.2845096 -0.1262364 -1.3682783
#> 6 -1.7690202 -2.7114759  4.652995344 -1.5605501  1.2200321 -0.9514023
#>          PC8        PC9       PC10
#> 1  1.3633561 -1.4703846  3.5072811
#> 2 -1.2569230  2.4778511  0.5030242
#> 3  3.9990525 -3.9571915  0.1248838
#> 4  4.3477501 -4.4694407  0.8116271
#> 5 -3.9339657 -0.0664483 -1.3033802
#> 6  0.1844682  1.5659617  3.6812083

## Example 2: Automatically determine the number of PCs to retain based on the variance threshold
filtered_covariates <- coVar_Filter(genotype_matrix, covariate_matrix, num_PCs = "varPCs", variance_threshold = 0.80)